requirements.txt
Flask
Flask-Cors
selenium
webdriver-manager
firebase-admin
twilio
python-dotenv